Run test mux on push
This GitHub Actions workflow creates a dummy video and attempts a mux on
each episode. By doing a test mux on each push, issues such as missing
fonts, invalid SubKt configs, negative times, and other issues that
would cause SubKt to fail will surface as soon as possible.

The .github directory is named .github.sample to avoid having GitHub
recognize this file as a workflow to apply on _this_ repo.
The intention is to have the user rename the directory when setting up
a new project.

This workflow is adapted from the workflow used in ButterFS' Hataraku Saibou repo:
https://github.com/butterfansubs/hataraku-saibou/blob/8ec814aa796539c4838faad1e3cf7c97f223b99a/.github/workflows/main-s2.yaml