build: `File()` is not guaranteed to work, `file()` is
gradle fuckery with relative paths and cwd:

Never use `new File(relative path)` unless passed to `file()` or `files()` or `from()`
or other methods defined in terms of `file()` or `files()`.
Otherwise, this creates a path relative to the current working directory (CWD).
Gradle can make no guarantees about the location of the CWD,
which means builds that rely on it may break at any time.

Co-Authored-By: petzku <5639335+petzku@users.noreply.github.com>